Added Verifikasi Agenda

+++ resources/views/agenda/index.blade.php
@@ -5,10 +5,12 @@
 
 
 @section('content')
-    <h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">Tables /</span> Basic Tables</h4>
+    <h4 class="fw-bold py-3 mb-4"><span class="text-muted fw-light">Permohonan /</span> Agenda</h4>
     <div class="card">
-        <h5 class="card-header">Table Basic</h5>
-        <a href="agenda/create" class="btn btn-primary">Create</a>
+        <h5 class="card-header">Daftar Agenda</h5>
+        @if (auth()->user()->id === 0)
+            <a href="agenda/create" class="btn btn-primary">Create</a>
+        @endif
         <div class="table-responsive text-nowrap">
             <table class="table">
                 <thead>
@@ -43,27 +45,28 @@
                                         <i class="bx bx-dots-vertical-rounded"></i>
                                     </button>
                                     <div class="dropdown-menu" style="">
-                                        <a class="dropdown-item" href="{{ route('agenda.edit', ['id' => $agenda->id]) }}">
-                                            <i class="bx bx-edit-alt me-1"></i> Edit
-                                        </a>
-
                                         <a class="dropdown-item" href="{{ route('agenda.detail', ['id' => $agenda->id]) }}">
                                             <i class="bx bx-detail me-1"></i>Detail
                                         </a>
+                                        @if (auth()->user()->id === 0)
+                                            <a class="dropdown-item"
+                                                href="{{ route('agenda.edit', ['id' => $agenda->id]) }}">
+                                                <i class="bx bx-edit-alt me-1"></i> Edit
+                                            </a>
 
-                                        <a class="dropdown-item" href="javascript:void(0);"
-                                            onclick="deleteAgenda({{ $agenda->id }})">
-                                            <i class="bx bx-trash me-1"></i> Delete
-                                        </a>
-
+                                            <a class="dropdown-item" href="javascript:void(0);"
+                                                onclick="deleteAgenda({{ $agenda->id }})">
+                                                <i class="bx bx-trash me-1"></i> Delete
+                                            </a>
                                     </div>
-                                </div>
-                            </td>
-                        </tr>
-                    @endforeach
-                </tbody>
-            </table>
+                    @endif
         </div>
+        </td>
+        </tr>
+        @endforeach
+        </tbody>
+        </table>
+    </div>
     </div>
 
     <script>
